+++ secrets/secrets.nix
@@ -16,6 +16,11 @@ in {
     inherit (hosts) phi sigma;
   };
 
+  "x11vnc_phi.age".publicKeys = builtins.attrValues {
+    inherit (users) enzime_phi;
+    inherit (hosts) phi;
+  };
+
   "zshrc_phi.age".publicKeys = builtins.attrValues {
     inherit (users) enzime_phi;
     inherit (hosts) phi;